+++ core/config.h
@@ -326,14 +326,14 @@ namespace RayGene3D
 
   public:
     virtual const std::shared_ptr<Batch>& CreateBatch(const std::string& name,
-      const std::pair<const Batch::Entity*, uint32_t>& entities,
-      const std::pair<const Batch::Sampler*, uint32_t>& samplers = {},
-      const std::pair<const std::shared_ptr<View>*, uint32_t>& ub_views = {},
-      const std::pair<const std::shared_ptr<View>*, uint32_t>& sb_views = {},
-      const std::pair<const std::shared_ptr<View>*, uint32_t>& ri_views = {},
-      const std::pair<const std::shared_ptr<View>*, uint32_t>& wi_views = {},
-      const std::pair<const std::shared_ptr<View>*, uint32_t>& rb_views = {},
-      const std::pair<const std::shared_ptr<View>*, uint32_t>& wb_views = {}
+      const std::pair<const Batch::Entity*, size_t>& entities,
+      const std::pair<const Batch::Sampler*, size_t>& samplers = {},
+      const std::pair<const std::shared_ptr<View>*, size_t>& ub_views = {},
+      const std::pair<const std::shared_ptr<View>*, size_t>& sb_views = {},
+      const std::pair<const std::shared_ptr<View>*, size_t>& ri_views = {},
+      const std::pair<const std::shared_ptr<View>*, size_t>& wi_views = {},
+      const std::pair<const std::shared_ptr<View>*, size_t>& rb_views = {},
+      const std::pair<const std::shared_ptr<View>*, size_t>& wb_views = {}
     ) = 0;
     //void VisitBatch(std::function<void(const std::shared_ptr<Batch>&)> visitor) { for (const auto& batch : batches) visitor(batch); }
     void DestroyBatch(const std::shared_ptr<Batch>& batch) 
@@ -357,7 +357,7 @@ namespace RayGene3D
       Pass& pass,
       const std::string& source,
       Config::Compilation compilation,
-      const std::pair<const std::pair<std::string, std::string>*, uint32_t>& defines,
+      const std::pair<const std::pair<std::string, std::string>*, size_t>& defines,
       const Config::IAState& ia_state,
       const Config::RCState& rc_state,
       const Config::DSState& ds_state,
@@ -366,7 +366,7 @@ namespace RayGene3D
       Pass& pass,
       const std::string& source,
       Config::Compilation compilation,
-      const std::pair<const std::pair<std::string, std::string>*, uint32_t>& defines,
+      const std::pair<const std::pair<std::string, std::string>*, size_t>& defines,
       const Config::RCState& rc_state,
       const Config::DSState& ds_state,
       const Config::OMState& om_state);
@@ -374,7 +374,7 @@ namespace RayGene3D
       Pass& pass,
       const std::string& source,
       Config::Compilation compilation,
-      const std::pair<const std::pair<std::string, std::string>*, uint32_t>& defines);
+      const std::pair<const std::pair<std::string, std::string>*, size_t>& defines);
     virtual ~Config();
   };
