+++ core/vlk/vlk_config.cpp
@@ -883,7 +883,7 @@ namespace RayGene3D
     Pass& pass,
     const std::string& source,
     Config::Compilation compilation,
-    const std::pair<const std::pair<std::string, std::string>*, uint32_t>& defines,
+    const std::pair<const std::pair<std::string, std::string>*, size_t>& defines,
     const Config::IAState& ia_state,
     const Config::RCState& rc_state,
     const Config::DSState& ds_state,
@@ -897,7 +897,7 @@ namespace RayGene3D
     Pass& pass,
     const std::string& source,
     Config::Compilation compilation,
-    const std::pair<const std::pair<std::string, std::string>*, uint32_t>& defines,
+    const std::pair<const std::pair<std::string, std::string>*, size_t>& defines,
     const Config::RCState& rc_state,
     const Config::DSState& ds_state,
     const Config::OMState& om_state)
@@ -910,7 +910,7 @@ namespace RayGene3D
     Pass& pass,
     const std::string& source,
     Config::Compilation compilation,
-    const std::pair<const std::pair<std::string, std::string>*, uint32_t>& defines)
+    const std::pair<const std::pair<std::string, std::string>*, size_t>& defines)
     : Config(name, pass, source, compilation, defines)
   {
     VLKConfig::Initialize();
